评估

+++ src/com/example/demo/User.java
@@ -6,7 +6,7 @@ public class User {
 	private String uname;
 	private String upassword;
 	private int uage;
-	private int ulevel;
+	private double ulevel;
 	private String usex;
 	private String uposition;
 	private String urealname;
@@ -40,10 +40,10 @@ public class User {
 	public int getUage(){
 		return uage;
 	}
-	public void setUlevel(int ulevel){
+	public void setUlevel(double ulevel){
 		this.ulevel=ulevel;
 	}
-	public int getUlevel(){
+	public double getUlevel(){
 		return ulevel;
 	}
 	public void setUsex(String usex){
